Corrigé le test suite à l'intégration de la dernière version du web-service Party 3 qui restreint la plage de validité des numéros de contribuables (conversion long -> int sur les numéros de contribuable).
git-svn-id: https://svn.etat-de-vaud.ch/SVN_SOCLE/registre/unireg/trunk@62643 cb3efa0a-b8b0-9f43-a4c8-343d8b7e4362

+++ unireg/ws/src/test/java/ch/vd/uniregctb/webservices/tiers3/cache/PartyWebServiceCacheTest.java
@@ -591,7 +591,7 @@ public class PartyWebServiceCacheTest extends WebserviceTest {
 			// on vérifie qu'il y a un second appel au web-service, mais qu'il ne concerne que le tiers Eric
 			assertEquals(2, implementation.getBatchTiersCalls.size());
 			assertEquals(params.getPartyNumbers(), implementation.getBatchTiersCalls.get(0).getPartyNumbers());
-			assertEquals(Arrays.asList(ids.eric), implementation.getBatchTiersCalls.get(1).getPartyNumbers());
+			assertEquals(Arrays.asList(ids.eric.intValue()), implementation.getBatchTiersCalls.get(1).getPartyNumbers());
 		}
 	}
